# swagger.json
{"openapi":"3.0.0","paths":{"/yahoo-finance/historical":{"get":{"operationId":"YahooFinanceController_getHistoricalData","summary":"Get historical data for a symbol","description":"Retrieves historical price and volume data for a given stock symbol","parameters":[{"name":"symbol","required":true,"in":"query","description":"Stock symbol (e.g., AAPL for Apple Inc.)","example":"AAPL","schema":{"type":"string"}},{"name":"from","required":true,"in":"query","description":"Start date (YYYY-MM-DD)","example":"2023-01-01","schema":{"type":"string"}},{"name":"to","required":true,"in":"query","description":"End date (YYYY-MM-DD)","example":"2023-12-31","schema":{"type":"string"}},{"name":"interval","required":false,"in":"query","description":"Data interval","schema":{"enum":["1d","1wk","1mo"],"type":"string"}}],"responses":{"200":{"description":"Historical data retrieved successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HistoricalDataDTO"}}}}},"404":{"description":"Historical data not found"},"500":{"description":"Internal server error"}},"tags":["Yahoo Finance"]}},"/yahoo-finance/insights":{"get":{"operationId":"YahooFinanceController_getInsights","summary":"Get insights for a symbol","description":"Retrieves insights and analysis for a given stock symbol","parameters":[{"name":"symbol","required":true,"in":"query","description":"Stock symbol (e.g., AAPL for Apple Inc.)","example":"AAPL","schema":{"type":"string"}},{"name":"reportsCount","required":false,"in":"query","description":"Number of reports to retrieve","schema":{"type":"number"}}],"responses":{"200":{"description":"Insights retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InsightsDTO"}}}},"404":{"description":"Insights not found"},"500":{"description":"Internal server error"}},"tags":["Yahoo Finance"]}},"/yahoo-finance/quote":{"get":{"operationId":"YahooFinanceController_getQuote","summary":"Get quote for a symbol","description":"Retrieves current market data for a given stock symbol","parameters":[{"name":"symbol","required":true,"in":"query","description":"Stock symbol (e.g., AAPL for Apple Inc.)","example":"AAPL","schema":{"type":"string"}}],"responses":{"200":{"description":"Quote retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuoteDTO"}}}},"404":{"description":"Quote not found"},"500":{"description":"Internal server error"}},"tags":["Yahoo Finance"]}},"/yahoo-finance/quote-summary":{"get":{"operationId":"YahooFinanceController_getQuoteSummary","summary":"Get quote summary for a symbol","description":"Retrieves a comprehensive summary of data for a given stock symbol","parameters":[{"name":"symbol","required":true,"in":"query","description":"Stock symbol (e.g., AAPL for Apple Inc.)","example":"AAPL","schema":{"type":"string"}},{"name":"modules","required":true,"in":"query","description":"Comma-separated list of data modules to retrieve","example":"assetProfile,financialData","schema":{"type":"string"}}],"responses":{"200":{"description":"Quote summary retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuoteSummaryDTO"}}}},"404":{"description":"Quote summary not found"},"500":{"description":"Internal server error"}},"tags":["Yahoo Finance"]}},"/yahoo-finance/search":{"get":{"operationId":"YahooFinanceController_search","summary":"Search for symbols","description":"Searches for stock symbols based on a query string","parameters":[{"name":"query","required":true,"in":"query","description":"Search query","example":"Apple","schema":{"type":"string"}},{"name":"quotesCount","required":false,"in":"query","description":"Number of quotes to retrieve","schema":{"type":"number"}},{"name":"newsCount","required":false,"in":"query","description":"Number of news items to retrieve","schema":{"type":"number"}}],"responses":{"200":{"description":"Search results retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchResultDTO"}}}},"404":{"description":"No results found"},"500":{"description":"Internal server error"}},"tags":["Yahoo Finance"]}},"/yahoo-finance/chart":{"get":{"operationId":"YahooFinanceController_getChart","summary":"Get chart data for a symbol","description":"Retrieves chart data for a given stock symbol","parameters":[{"name":"symbol","required":true,"in":"query","description":"Stock symbol (e.g., AAPL for Apple Inc.)","example":"AAPL","schema":{"type":"string"}},{"name":"from","required":true,"in":"query","description":"Start date (YYYY-MM-DD)","example":"2023-01-01","schema":{"type":"string"}},{"name":"to","required":true,"in":"query","description":"End date (YYYY-MM-DD)","example":"2023-12-31","schema":{"type":"string"}},{"name":"interval","required":false,"in":"query","description":"Data interval","schema":{"type":"string"}}],"responses":{"200":{"description":"Chart data retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChartResultDTO"}}}},"404":{"description":"Chart data not found"},"500":{"description":"Internal server error"}},"tags":["Yahoo Finance"]}},"/yahoo-finance/trending-symbols":{"get":{"operationId":"YahooFinanceController_getTrendingSymbols","summary":"Get trending symbols","description":"Retrieves trending stock symbols for a given country","parameters":[{"name":"country","required":true,"in":"query","description":"Country code (e.g., US for United States)","example":"US","schema":{"type":"string"}},{"name":"count","required":false,"in":"query","description":"Number of trending symbols to retrieve","schema":{"type":"number"}}],"responses":{"200":{"description":"Trending symbols retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrendingSymbolsDTO"}}}},"404":{"description":"Trending symbols not found"},"500":{"description":"Internal server error"}},"tags":["Yahoo Finance"]}},"/yahoo-finance/recommendations":{"get":{"operationId":"YahooFinanceController_getRecommendations","summary":"Get recommendations for a symbol","description":"Retrieves stock recommendations based on a given symbol","parameters":[{"name":"symbol","required":true,"in":"query","description":"Stock symbol (e.g., AAPL for Apple Inc.)","example":"AAPL","schema":{"type":"string"}}],"responses":{"200":{"description":"Recommendations retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecommendationDTO"}}}},"404":{"description":"Recommendations not found"},"500":{"description":"Internal server error"}},"tags":["Yahoo Finance"]}},"/yahoo-finance/options":{"get":{"operationId":"YahooFinanceController_getOptions","summary":"Get options data for a symbol","description":"Retrieves options data for a given stock symbol","parameters":[{"name":"symbol","required":true,"in":"query","description":"Stock symbol (e.g., AAPL for Apple Inc.)","example":"AAPL","schema":{"type":"string"}}],"responses":{"200":{"description":"Options data retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OptionsDTO"}}}},"404":{"description":"Options data not found"},"500":{"description":"Internal server error"}},"tags":["Yahoo Finance"]}},"/yahoo-finance/stock-image":{"get":{"operationId":"YahooFinanceController_getStockImage","summary":"Get stock image for a symbol","description":"Retrieves the stock image URL for a given stock symbol","parameters":[{"name":"symbol","required":true,"in":"query","description":"Stock symbol (e.g., AAPL for Apple Inc.)","example":"AAPL","schema":{"type":"string"}}],"responses":{"200":{"description":"Stock image URL retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StockImageDTO"}}}},"404":{"description":"Stock image not found"},"500":{"description":"Internal server error"}},"tags":["Yahoo Finance"]}}},"info":{"title":"Stonks - Yahoo Finance API","description":"API for fetching Yahoo Finance data through the Stonks service","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"HistoricalDataDTO":{"type":"object","properties":{"date":{"type":"string","description":"The date of the data point","example":"2023-01-01"},"open":{"type":"number","description":"The opening price","example":150.23},"high":{"type":"number","description":"The highest price of the day","example":152.85},"low":{"type":"number","description":"The lowest price of the day","example":149.56},"close":{"type":"number","description":"The closing price","example":151.73},"adjClose":{"type":"number","description":"The adjusted closing price","example":151.73},"volume":{"type":"number","description":"The trading volume","example":1000000}},"required":["date","open","high","low","close","adjClose","volume"]},"InsightsDTO":{"type":"object","properties":{"symbol":{"type":"string","description":"The stock symbol","example":"AAPL"},"instrumentInfo":{"type":"object","description":"Technical analysis insights"},"companySnapshot":{"type":"object","description":"Company snapshot information"},"recommendation":{"type":"object","description":"Analyst recommendations"}},"required":["symbol","instrumentInfo","companySnapshot","recommendation"]},"QuoteDTO":{"type":"object","properties":{"symbol":{"type":"string","description":"The stock symbol","example":"AAPL"},"regularMarketPrice":{"type":"number","description":"The current market price","example":150.25},"regularMarketChange":{"type":"number","description":"The price change","example":2.5},"regularMarketChangePercent":{"type":"number","description":"The price change percentage","example":1.69},"marketCap":{"type":"number","description":"The market cap","example":2500000000000},"regularMarketVolume":{"type":"number","description":"The trading volume","example":100000000},"fiftyTwoWeekHigh":{"type":"number","description":"The 52-week high","example":180.5},"fiftyTwoWeekLow":{"type":"number","description":"The 52-week low","example":120.75}},"required":["symbol","regularMarketPrice","regularMarketChange","regularMarketChangePercent","marketCap","regularMarketVolume","fiftyTwoWeekHigh","fiftyTwoWeekLow"]},"QuoteSummaryDTO":{"type":"object","properties":{"assetProfile":{"type":"object","description":"Company profile information"},"financialData":{"type":"object","description":"Financial data"}},"required":["assetProfile","financialData"]},"SearchResultDTO":{"type":"object","properties":{"quotes":{"description":"List of matching quotes","type":"array","items":{"$ref":"#/components/schemas/QuoteDTO"}},"news":{"description":"List of related news items","type":"array","items":{"type":"string"}}},"required":["quotes","news"]},"ChartResultDTO":{"type":"object","properties":{"meta":{"type":"object","description":"Meta information about the chart"},"timestamp":{"description":"Timestamp array for the chart data points","type":"array","items":{"type":"number"}},"indicators":{"type":"object","description":"Indicators data for the chart"}},"required":["meta","timestamp","indicators"]},"TrendingSymbolsDTO":{"type":"object","properties":{"count":{"type":"number","description":"Count of trending symbols","example":5},"quotes":{"description":"List of trending symbols","type":"array","items":{"type":"string"}},"jobTimestamp":{"type":"number","description":"Timestamp of the job","example":1625097600000},"startInterval":{"type":"number","description":"Start interval","example":1625097600000}},"required":["count","quotes","jobTimestamp","startInterval"]},"RecommendationDTO":{"type":"object","properties":{"symbol":{"type":"string","description":"The stock symbol","example":"AAPL"},"recommendedSymbols":{"description":"List of recommended symbols","type":"array","items":{"type":"string"}}},"required":["symbol","recommendedSymbols"]},"OptionsDTO":{"type":"object","properties":{"underlyingSymbol":{"type":"string","description":"The underlying stock symbol","example":"AAPL"},"expirationDates":{"description":"List of expiration dates","type":"array","items":{"format":"date-time","type":"string"}},"strikes":{"description":"List of available strikes","type":"array","items":{"type":"number"}},"hasMiniOptions":{"type":"boolean","description":"Whether the stock has mini options","example":false},"quote":{"description":"Quote data for the underlying stock","allOf":[{"$ref":"#/components/schemas/QuoteDTO"}]},"options":{"description":"List of options data","type":"array","items":{"type":"string"}}},"required":["underlyingSymbol","expirationDates","strikes","hasMiniOptions","quote","options"]},"StockImageDTO":{"type":"object","properties":{"imageUrl":{"type":"string","description":"URL of the stock image","example":"https://example.com/stock-images/AAPL.png"}},"required":["imageUrl"]}}}}